Repair Recycler View Expense Income, Update and Delete Activity (still not using MVVM)

+++ app/src/main/java/edu/bluejack20_1/dearmory/repositories/ExpenseIncomeRepository.kt
@@ -40,7 +40,7 @@ class ExpenseIncomeRepository private constructor(){
                         .setId(data.child("id").getValue().toString())
                         .setNotes(data.child("notes").getValue().toString())
                         .setTime(data.child("time").getValue().toString())
-                        .setAmount(data.child("amount").getValue().toString().toInt()))
+                        .setAmount(data.child("amount").getValue().toString().toLong()))
                 }
                 expenseIncomeLiveData.postValue(expenseIncomeModels)
             }
